feat/security/#168 : 토큰 오류 수정하였습니다.
