+++ src/main/java/com/ghbt/ghbt_starbucks/api/auth/controller/AuthController.java
@@ -6,19 +6,14 @@ import com.ghbt.ghbt_starbucks.global.security.dto.SignupDto;
 import com.ghbt.ghbt_starbucks.global.security.dto.TokenDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpCookie;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.web.bind.annotation.CookieValue;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
@@ -80,33 +75,5 @@ public class AuthController {
         return ResponseEntity.status(HttpStatus.OK)
             .header(HttpHeaders.SET_COOKIE, responseCookie.toString())
             .build();
-    }
-
-    /**
-     * 토큰 재발급
-     */
-    @PostMapping("/reissue")
-    @Operation(summary = "토큰 재발급 API", description = "상세 기능 : AccessToken 이 만료되었을 때 재발급해주는 API 입니다.")
-    public ResponseEntity<?> reissue(@CookieValue(name = "refresh-token") String refreshToken) {
-
-        TokenDto reissuedTokenDto = authService.reissue(refreshToken);
-        if (reissuedTokenDto != null) {
-            return ResponseEntity.status(HttpStatus.OK)
-                .header(HttpHeaders.SET_COOKIE, "refresh-token=" + reissuedTokenDto.getRefreshToken()
-                    + "; domain= localhost; path=/; SameSite=None; Secure; httpOnly;")
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + reissuedTokenDto.getAccessToken())
-                .build();
-        } else {
-            ResponseCookie responseCookie = ResponseCookie.from("refresh-token", "")
-                .maxAge(0)
-                .path("/")
-                .build();
-            return ResponseEntity
-                .status(HttpStatus.UNAUTHORIZED)
-                .header(HttpHeaders.SET_COOKIE, "refresh-token=" + ""
-                    + "; domain= localhost; Max-Age=0; path=/; SameSite=None; Secure; httpOnly;")
-                .header(HttpHeaders.SET_COOKIE, responseCookie.toString())
-                .build();
-        }
     }
 }
